requirements.txt
PyYAML==5.3.1
spacy==2.3.5
